api: some refactoring
