+++ api/server.go
@@ -3,20 +3,16 @@ package api
 import (
 	"context"
 	"errors"
-	"io"
 	"io/fs"
 	"net"
 	"net/http"
-	"path"
-	"strings"
 
 	"github.com/gorilla/mux"
 	"github.com/rs/zerolog"
 	"golang.org/x/sync/errgroup"
-
-	"github.com/profclems/compozify/api/router"
 )
 
+// Server is the web app server.
 type Server struct {
 	logger   *zerolog.Logger
 	listener net.Listener
@@ -24,14 +20,16 @@ type Server struct {
 	assets   fs.FS
 }
 
+// NewServer creates a new Server.
 func NewServer(logger *zerolog.Logger, listener net.Listener, assets fs.FS) *Server {
 	server := &Server{
 		logger:   logger,
 		listener: listener,
 		assets:   assets,
 	}
+
 	r := mux.NewRouter()
-	router.Handle(r)
+	r.HandleFunc("/api/parse", server.ParseDockerCommand).Methods("POST")
 
 	//r.PathPrefix("/static").HandlerFunc(server.cacheHandler)
 	r.PathPrefix("/_next").HandlerFunc(server.cacheHandler)
@@ -42,39 +40,6 @@ func NewServer(logger *zerolog.Logger, listener net.Listener, assets fs.FS) *Ser
 	}
 
 	return server
-}
-
-// appHandler is web app http handler function.
-func (server *Server) appHandler(w http.ResponseWriter, r *http.Request) {
-	header := w.Header()
-
-	header.Set("Content-Type", "text/html; charset=UTF-8")
-	header.Set("X-Content-Type-Options", "nosniff")
-	header.Set("Referrer-Policy", "same-origin")
-
-	f, err := server.assets.Open("index.html")
-	if err != nil {
-		http.Error(w, `web/ unbuilt`, http.StatusNotFound)
-		return
-	}
-	defer func() { _ = f.Close() }()
-
-	_, _ = io.Copy(w, f)
-}
-
-func (server *Server) cacheHandler(w http.ResponseWriter, r *http.Request) {
-	staticServer := http.FileServer(http.FS(server.assets))
-	header := w.Header()
-
-	if contentType, ok := commonContentType(path.Ext(r.URL.Path)); ok {
-		header.Set("Content-Type", contentType)
-	}
-
-	header.Set("Cache-Control", "public, max-age=31536000")
-	header.Set("X-Content-Type-Options", "nosniff")
-	header.Set("Referrer-Policy", "same-origin")
-
-	staticServer.ServeHTTP(w, r)
 }
 
 // Run starts the server that host webapp and api endpoints.
@@ -101,32 +66,4 @@ func (server *Server) Run(ctx context.Context) (err error) {
 // Close closes server and underlying listener.
 func (server *Server) Close() error {
 	return server.http.Close()
-}
-
-func commonContentType(ext string) (string, bool) {
-	ext = strings.ToLower(ext)
-	mime, ok := commonTypes[ext]
-	return mime, ok
-}
-
-var commonTypes = map[string]string{
-	".css":   "text/css; charset=utf-8",
-	".gif":   "image/gif",
-	".htm":   "text/html; charset=utf-8",
-	".html":  "text/html; charset=utf-8",
-	".jpeg":  "image/jpeg",
-	".jpg":   "image/jpeg",
-	".js":    "application/javascript",
-	".mjs":   "application/javascript",
-	".otf":   "font/otf",
-	".pdf":   "application/pdf",
-	".png":   "image/png",
-	".svg":   "image/svg+xml",
-	".ttf":   "font/ttf",
-	".wasm":  "application/wasm",
-	".webp":  "image/webp",
-	".xml":   "text/xml; charset=utf-8",
-	".sfnt":  "font/sfnt",
-	".woff":  "font/woff",
-	".woff2": "font/woff2",
 }
